refactor: fix some code-smels
