+++ src/App.tsx
@@ -1,36 +1,10 @@
-// import MainPage from './pages/MainPage';
-// import ErrorBoundary from './components/ErrorBoundary';
-// import styles from './App.module.css';
-import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
-import Layout from './Layout';
-import CardDetails from './pages/CardDetails';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <Layout />,
-    children: [
-      {
-        path: '/details/:id',
-        element: <CardDetails />,
-      },
-      {
-        path: '*',
-        element: <Navigate to="/notfound" />,
-      },
-    ],
-  },
-]);
+import { RouterProvider } from 'react-router-dom';
+import router from './utils/router';
 
 function App() {
   return (
     <>
       <RouterProvider router={router} />
-      {/* <main className={styles.container}> */}
-      {/* <ErrorBoundary>
-            <MainPage />
-          </ErrorBoundary> */}
-      {/* </main> */}
     </>
   );
 }
