+++ src/Layout.tsx
@@ -1,5 +1,5 @@
 import ErrorBoundary from './components/ErrorBoundary';
-import RouteProvider from './components/RouteProvider';
+import NotFoundProvider from './components/NotFoundProvider';
 import SearchPage from './pages/SearchPage';
 import { Outlet } from 'react-router-dom';
 
@@ -7,12 +7,12 @@ function Layout() {
   return (
     <>
       <div className="wrapper">
-        <RouteProvider>
+        <NotFoundProvider>
           <ErrorBoundary>
             <SearchPage />
             <Outlet />
           </ErrorBoundary>
-        </RouteProvider>
+        </NotFoundProvider>
       </div>
     </>
   );
