+++ src/__tests__/Pagination.test.tsx
@@ -1,0 +1,8 @@
+// import { screen } from "@testing-library/react"
+// import Pagination from '../components/Pagination';
+
+// describe("Pagination Component", () => {
+//   it('is 1 page', () => {
+//     screen.
+//   })
+// })
