+++ src/utils/constats.ts
@@ -1,0 +1,1 @@
+export const allPages = 47;
